retour a f64

+++ src/commands/math.rs
@@ -1,6 +1,6 @@
 command!(multiply(_ctx, msg, args) {
-    let one = args.single::<i128>().unwrap();
-    let two = args.single::<i128>().unwrap();
+    let one = args.single::<f64>().unwrap();
+    let two = args.single::<f64>().unwrap();
 
     let product = one * two;
